created prayers route for the function getAll
