+++ server/routers/prayersRouter.js
@@ -1,34 +1,18 @@
-// import the knex database connection 
-const db = require('../data/db-config.js');
+const express = require('express');
 
-// export model functions
-module.exports = {
-    getAll,
-    add,
-    update,
-    remove
-}
+const Prayers = require('./prayersModel.js')
 
-// return a list of all prayers with people ID
-function getAll(id) {
-    return db('prayers').where({'person_id': id});
-}
+const router = express.Router();
 
-// insert a new prayer into the database
-function add(prayer) {
-    return db('prayers').insert(prayer)
-        .then(ids => {
-            const id = ids[0];
-            return db('prayers').where({id}).first();
+// root route returns a list of all the prayers for the person
+router.get('/:id', (req, res) => {
+    const { id } = req.params;
+
+    Prayers.getAll(id)
+        .then(prayers => {
+            res.status(200).json(prayers);
+        })
+        .catch(err => {
+            res.status(500).json({message: "Failed to get a list of the prayers"});
         });
-}
-
-// update a prayer
-function update(id, changes) {
-    return db('prayers').where({ id }).update(changes);
-}
-
-// delete prayer by id
-function remove(id) {
-    return db('prayers').where({ id }).del();
-}
+});
